fix: mypage UI
