+++ week4/src/pages/myPage/MyPage.tsx
@@ -1,5 +1,22 @@
+import * as S from './MyPage.style';
+
 const MyPage = () => {
-  return <div>Mypage</div>;
+  return (
+    <S.SignUpWrapper>
+      <S.SignUpTitle> MY PAGE</S.SignUpTitle>
+      <S.SignUpContainer>
+        <S.SignUpImage />
+        <S.UserInfoWrapper>
+          <S.UserInfoContainer>
+            <S.UserInfoText>ID: </S.UserInfoText> <S.UserInfoText> </S.UserInfoText>
+          </S.UserInfoContainer>
+          <S.UserInfoContainer>
+            <S.UserInfoText>닉네임: </S.UserInfoText> <S.UserInfoText> </S.UserInfoText>
+          </S.UserInfoContainer>
+        </S.UserInfoWrapper>
+      </S.SignUpContainer>
+    </S.SignUpWrapper>
+  );
 };
 
 export default MyPage;
